Fix millis() and micros() rollover bug
Also replace long with int32_t, so it is not architecture dependent

+++ BTD.cpp
@@ -384,7 +384,7 @@ uint8_t BTD::Release() {
 uint8_t BTD::Poll() {
         if(!bPollEnable)
                 return 0;
-        if((long)(millis() - qNextPollTime) >= 0L) { // Don't poll if shorter than polling interval
+        if((int32_t)(millis() - qNextPollTime) >= 0L) { // Don't poll if shorter than polling interval
                 qNextPollTime = millis() + pollInterval; // Set new poll time
                 HCI_event_task(); // Poll the HCI event pipe
                 HCI_task(); // HCI state machine
