+++ examples/testusbhostFAT/testusbhostFAT.ino
@@ -371,7 +371,7 @@ void serialEvent() {
 // ALL teensy versions LACK PWM ON LED
 
 ISR(TIMER3_COMPA_vect) {
-        if((long)(millis() - LEDnext_time) >= 0L) {
+        if((int32_t)(millis() - LEDnext_time) >= 0L) {
                 LEDnext_time = millis() + 30;
 
                 // set the brightness of LED
@@ -407,7 +407,7 @@ void loop() {
 
 #if defined(__AVR__)
         // Print a heap status report about every 10 seconds.
-        if((long)(millis() - HEAPnext_time) >= 0L) {
+        if((int32_t)(millis() - HEAPnext_time) >= 0L) {
                 if(UsbDEBUGlvl > 0x50) {
                         printf_P(PSTR("Available heap: %u Bytes\r\n"), freeHeap());
                 }
@@ -421,7 +421,7 @@ void loop() {
 #endif
         // Horrid! This sort of thing really belongs in an ISR, not here!
         // We also will be needing to test each hub port, we don't do this yet!
-        if(!change && !usbon && (long)(millis() - usbon_time) >= 0L) {
+        if(!change && !usbon && (int32_t)(millis() - usbon_time) >= 0L) {
                 change = true;
                 usbon = true;
         }
